[XrdCl] Keep TPC plug-in support out of the public ABI
The third party copy support added a virtual method to
FileSystemPlugIn in the installed XrdClPlugInInterface.hh. That
changes the vtable layout of a class out-of-tree client plug-ins
derive from, which is an incompatible ABI change and would
require a major version bump. FileSystem::ThirdPartyCopy and
ProgressHandler were compatible additions, but froze a
first-iteration synchronous API in installed headers.

Restore the three installed headers to their released state and
move the whole surface to a new non-installed header instead:
a ThirdPartyCopyPlugIn capability interface which plug-in
filesystems inherit in addition to FileSystemPlugIn, and a free
function which looks up the plug-in factory for the source URL
and discovers the capability with a dynamic_cast. The interface
destructor is defined out of line so its typeinfo is anchored in
libXrdCl and the dynamic_cast works on objects from dlopened
plug-in libraries.

This also fixes the streaming fallback for plug-ins without TPC
support: the default virtual implementation returned
errNotImplemented, which the fall back copy job does not treat
as a reason to stream; the free function reports errNotSupported
when the capability is absent.
