+++ src/XrdClHttp/XrdClHttpFilesystem.hh
@@ -27,6 +27,7 @@
 #include <XrdCl/XrdClFileSystem.hh>
 #include <XrdCl/XrdClLog.hh>
 #include <XrdCl/XrdClPlugInInterface.hh>
+#include <XrdCl/XrdClThirdPartyCopyPlugIn.hh>
 #include <XrdCl/XrdClURL.hh>
 
 #include <shared_mutex>
@@ -48,7 +49,8 @@ namespace XrdClHttp {
 class HandlerQueue;
 class CurlOperation;
 
-class Filesystem final : public XrdCl::FileSystemPlugIn {
+class Filesystem final : public XrdCl::FileSystemPlugIn,
+                         public XrdCl::ThirdPartyCopyPlugIn {
 public:
     Filesystem(const std::string &, std::shared_ptr<HandlerQueue> queue, XrdCl::Log *log);
 
@@ -99,13 +101,13 @@ public:
                                       const XrdCl::Buffer     &arg,
                                       XrdCl::ResponseHandler  *handler,
                                       time_t                   timeout) override;
-                                      
+
     virtual XrdCl::XRootDStatus ThirdPartyCopy( const std::string            &source,
                                                 const std::string            &dest,
                                                 const XrdCl::PropertyList    *properties,
                                                 XrdCl::ProgressHandler       *progress_handler,
                                                 time_t                        timeout = 0 ) override;
-    
+
 private:
     XrdCl::XRootDStatus QueueOperation(
         std::unique_ptr<CurlOperation> operation,
